Add product scope info to the rest api
